don't type cast spaces remaining as int because INF is a float and casting to int makes it 0; [touch:10638]

+++ core/db_classes/EE_Event.class.php
@@ -896,7 +896,7 @@ class EE_Event extends EE_CPT_Base implements EEI_Line_Item_Object, EEI_Admin_Li
             }
         }
         return $filtered
-            ? (int) apply_filters(
+            ? apply_filters(
                 'FHEE_EE_Event__spaces_remaining',
                 $spaces_remaining,
                 $this,
